+++ src/alike.c
@@ -18,6 +18,8 @@ Go to <https://www.r-project.org/Licenses/GPL-2> for a copy of the license.
 
 #include "settings.h"
 #include "alike.h"
+#include "backports.h" // for R_getVar*
+#include "validate.h"  // For VALC_UnboundValue, should really merge headers
 
 /*-----------------------------------------------------------------------------\
 \-----------------------------------------------------------------------------*/
@@ -493,8 +495,9 @@ struct ALIKEC_res ALIKEC_alike_rec(
           for(i = 0; i < tar_len; i++) {
             const char * var_name_chr = CHAR(STRING_ELT(tar_names, i));
             SEXP var_name = PROTECT(install(var_name_chr));
-            SEXP var_cur_val = PROTECT(findVarInFrame(current, var_name));
-            if(var_cur_val == R_UnboundValue) {
+            SEXP var_cur_val =
+                PROTECT(R_getVarEx(var_name, current, false, VALC_UnboundValue));
+            if(var_cur_val == VALC_UnboundValue) {
               REPROTECT(res.wrap = allocVector(VECSXP, 2), ipx);
               res.success = 0;
               res.dat.strings.tar_pre = "contain";
@@ -503,21 +506,14 @@ struct ALIKEC_res ALIKEC_alike_rec(
               res.dat.strings.cur_pre = "";
               res.dat.strings.current[1] = ""; // gcc-10
             } else {
-              SEXP var_in_frame = PROTECT(findVarInFrame(target, var_name));
-              // Could find a promise, which we should evaluate. It shouldn't
-              // matter where we evaluate it, so eval in empty env.
-              if(TYPEOF(var_in_frame) == PROMSXP) {
-                var_in_frame = PROTECT(eval(var_in_frame, R_EmptyEnv));
-              } else PROTECT(R_NilValue); // Stack balance
-              if(TYPEOF(var_cur_val) == PROMSXP) {
-                var_cur_val = PROTECT(eval(var_cur_val, R_EmptyEnv));
-              } else PROTECT(R_NilValue); // Stack balance
-
+              // Since we got the names earlier, the var must exist (could an
+              // active binding delete it? We're not goint to worry about that).
+              SEXP var_in_frame = PROTECT(R_getVar(var_name, target, false));
               res = ALIKEC_alike_rec(
                 var_in_frame, var_cur_val, res.dat.rec, set
               );
               REPROTECT(res.wrap, ipx);
-              UNPROTECT(3);
+              UNPROTECT(1);
               if(!res.success) {
                 res.dat.rec = ALIKEC_rec_ind_chr(res.dat.rec, var_name_chr);
             } }