+++ src/parse.c
@@ -17,6 +17,8 @@ Go to <https://www.r-project.org/Licenses/GPL-2> for a copy of the license.
 */
 
 #include "validate.h"
+// For R_getVar*
+#include "backports.h"
 
 /* -------------------------------------------------------------------------- *\
 \* -------------------------------------------------------------------------- */
@@ -170,7 +172,7 @@ SEXP VALC_sub_symbol(
       );
     }
     int var_found_resolves_symbol = 0;
-    if(findVar(lang, rho) != R_UnboundValue) {
+    if(R_getVarEx(lang, rho, true, VALC_UnboundValue) != VALC_UnboundValue) {
       SEXP found_val = PROTECT(eval(lang, rho));
       SEXPTYPE found_val_type = TYPEOF(found_val);
       if(found_val_type == LANGSXP || found_val_type == SYMSXP) {